Move log entry classes to clients module
