+++ java/clients/src/main/java/sleeper/clients/status/report/statestore/StateStoreCommitterLogEntry.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.dsl.statestore;
+package sleeper.clients.status.report.statestore;
 
 import java.time.Instant;
