+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/statestore/ReadStateStoreCommitterLogsTest.java
@@ -18,9 +18,9 @@ package sleeper.systemtest.drivers.statestore;
 import org.junit.jupiter.api.Test;
 import software.amazon.awssdk.services.cloudwatchlogs.model.ResultField;
 
-import sleeper.systemtest.dsl.statestore.StateStoreCommitSummary;
-import sleeper.systemtest.dsl.statestore.StateStoreCommitterRunFinished;
-import sleeper.systemtest.dsl.statestore.StateStoreCommitterRunStarted;
+import sleeper.clients.status.report.statestore.StateStoreCommitSummary;
+import sleeper.clients.status.report.statestore.StateStoreCommitterRunFinished;
+import sleeper.clients.status.report.statestore.StateStoreCommitterRunStarted;
 
 import java.time.Instant;
 import java.util.List;
